Fix incorrect extern declaration of poll(2).

+++ ndk/sources/android/libportable/arch-mips/poll.c
@@ -104,7 +104,7 @@ static inline short change_mips_events(short mips_events)
     return mips_events;
 }
 
-extern int poll(struct pollfd *, nfds_t, long);
+extern int poll(struct pollfd *, nfds_t, int);
 
 int WRAP(poll)(struct pollfd *fds, nfds_t nfds, long timeout)
 {
